MINOR move wrapper classes to appropriate package (#3472)
