+++ schema-rules/src/test/java/io/confluent/kafka/schemaregistry/rules/cel/CelExecutorIntegrationTest.java
@@ -40,8 +40,8 @@ import io.confluent.kafka.schemaregistry.storage.RuleSetHandler;
 import io.confluent.kafka.serializers.KafkaAvroDeserializer;
 import io.confluent.kafka.serializers.KafkaAvroDeserializerConfig;
 import io.confluent.kafka.serializers.KafkaAvroSerializer;
-import io.confluent.kafka.serializers.WrapperKeyDeserializer;
-import io.confluent.kafka.serializers.WrapperKeySerializer;
+import io.confluent.kafka.serializers.wrapper.WrapperKeyDeserializer;
+import io.confluent.kafka.serializers.wrapper.WrapperKeySerializer;
 import java.time.Duration;
 import java.util.AbstractMap.SimpleEntry;
 import java.util.ArrayList;
