+++ schema-serializer/src/main/java/io/confluent/kafka/serializers/wrapper/WrapperKeyDeserializer.java
@@ -14,48 +14,49 @@
  * limitations under the License.
  */
 
-package io.confluent.kafka.serializers;
+package io.confluent.kafka.serializers.wrapper;
 
+import io.confluent.kafka.serializers.AbstractKafkaSchemaSerDe;
 import java.util.Map;
 import org.apache.kafka.common.header.Headers;
-import org.apache.kafka.common.serialization.Serializer;
+import org.apache.kafka.common.serialization.Deserializer;
 
-public class WrapperKeySerializer<T> implements Serializer<T> {
+public class WrapperKeyDeserializer<T> implements Deserializer<T> {
 
-  private Serializer<T> inner;
+  private Deserializer<T> inner;
 
   /**
-   * Constructor used by Kafka producer.
+   * Constructor used by Kafka consumer.
    */
-  public WrapperKeySerializer() {
+  public WrapperKeyDeserializer() {
   }
 
   @Override
   public void configure(Map<String, ?> configs, boolean isKey) {
-    configure(new WrapperKeySerializerConfig(configs), isKey);
+    configure(new WrapperKeyDeserializerConfig(configs), isKey);
   }
 
   @SuppressWarnings("unchecked")
-  protected void configure(WrapperKeySerializerConfig config, boolean isKey) {
+  protected void configure(WrapperKeyDeserializerConfig config, boolean isKey) {
     if (!isKey) {
-      throw new IllegalArgumentException("WrapperKeySerializer is only for keys");
+      throw new IllegalArgumentException("WrapperKeyDeserializer is only for keys");
     }
     this.inner = config.getConfiguredInstance(
-        WrapperKeySerializerConfig.WRAPPED_KEY_SERIALIZER, Serializer.class);
+        WrapperKeyDeserializerConfig.WRAPPED_KEY_DESERIALIZER, Deserializer.class);
     this.inner.configure(config.originals(), isKey);
   }
 
   @Override
-  public byte[] serialize(String topic, T data) {
-    return serialize(topic, null, data);
+  public T deserialize(String topic, byte[] bytes) {
+    return deserialize(topic, null, bytes);
   }
 
   @Override
-  public byte[] serialize(String topic, Headers headers, T data) {
+  public T deserialize(String topic, Headers headers, byte[] bytes) {
     try {
-      return inner.serialize(topic, headers, data);
+      return inner.deserialize(topic, headers, bytes);
     } finally {
-      AbstractKafkaSchemaSerDe.setKey(data);
+      AbstractKafkaSchemaSerDe.setKey(bytes);
     }
   }
