+++ schema-serializer/src/main/java/io/confluent/kafka/serializers/wrapper/WrapperKeyDeserializerConfig.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2023 Confluent Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 3.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
@@ -14,31 +14,31 @@
  * limitations under the License.
  */
 
-package io.confluent.kafka.serializers;
+package io.confluent.kafka.serializers.wrapper;
 
 import java.util.Map;
 import org.apache.kafka.common.config.AbstractConfig;
 import org.apache.kafka.common.config.ConfigDef;
 
-public class WrapperKeySerializerConfig extends AbstractConfig {
+public class WrapperKeyDeserializerConfig extends AbstractConfig {
 
-  public static final String WRAPPED_KEY_SERIALIZER = "wrapped.key.serializer";
-  public static final String WRAPPED_KEY_SERIALIZER_DOC =
-      "Serializer class to which calls will be delegated";
+  public static final String WRAPPED_KEY_DESERIALIZER = "wrapped.key.deserializer";
+  public static final String WRAPPED_KEY_DESERIALIZER_DOC =
+      "Deserializer class to which calls will be delegated";
 
   private static final ConfigDef config;
 
   static {
     config = new ConfigDef().define(
-        WRAPPED_KEY_SERIALIZER,
+        WRAPPED_KEY_DESERIALIZER,
         ConfigDef.Type.CLASS,
         Object.class,
         ConfigDef.Importance.HIGH,
-        WRAPPED_KEY_SERIALIZER_DOC
+        WRAPPED_KEY_DESERIALIZER_DOC
     );
   }
 
-  public WrapperKeySerializerConfig(Map<?, ?> props) {
+  public WrapperKeyDeserializerConfig(Map<?, ?> props) {
     super(config, props);
   }
 }
